fix: replace custom switches with Tabs [MDS-340]
