+++ next-docs/components/facing/ForDevelopers.tsx
@@ -27,7 +27,7 @@ const ForDevelopers = () => {
         <GithubIcon />
         <h2 className="text-moon-24 grow 3xl:grow-0">For developers.</h2>
         <div className="absolute top-0 bottom-0 left-0 right-0 3xl:top-6 3xl:bottom-auto ltr:3xl:right-6 ltr:3xl:left-auto rtl:3xl:left-6 rtl:3xl:right-auto rtl:-rotate-90">
-          <Link href="/gettingStartedDevs">
+          <Link href="/gettingStarted">
             <a className="flex items-center justify-end h-full pe-6 3xl:pe-0 w-full rounded-moon-s-lg text-moon-32">
               <ArrowsTopRight />
             </a>
@@ -44,7 +44,7 @@ const ForDevelopers = () => {
         >
           Github
         </Button>
-        <Link href="/gettingStartedDevs" passHref>
+        <Link href="/gettingStarted" passHref>
           <Button as="a" variant="secondary">
             Getting Started
           </Button>
