--- next-docs/components/gettingStarted/DesDevSwitch.tsx
@@ -1,43 +1,0 @@
-import { mergeClassnames } from '@heathmont/moon-core-tw';
-import Link from 'next/link';
-
-const Button: React.FC<{ isActive: boolean; onClick: () => void }> = ({
-  children,
-  isActive,
-  ...rest
-}) => (
-  <a
-    className={mergeClassnames(
-      'group cursor-pointer transition-colors flex-inline rounded-md p-1.5 lg:ps-2.5 lg:pe-3.5 flex items-center text-sm font-medium shadow-sm',
-      isActive ? 'text-bulma bg-gohan' : 'text-trunks'
-    )}
-    {...rest}
-  >
-    {children}
-  </a>
-);
-
-type Props = {
-  isDesActive: boolean;
-  setDesActive: () => void;
-  setDevActive: () => void;
-};
-
-const DesDevSwitch = ({ isDesActive, setDesActive, setDevActive }: Props) => (
-  <div className="flex flex-row gap-2">
-    <div className="p-0.5 rounded-lg flex bg-beerus">
-      <Link href="/gettingStarted" passHref>
-        <Button isActive={isDesActive} onClick={setDesActive}>
-          I'm a designer
-        </Button>
-      </Link>
-      <Link href="/gettingStartedDevs" passHref>
-        <Button isActive={!isDesActive} onClick={setDevActive}>
-          I'm a developer
-        </Button>
-      </Link>
-    </div>
-  </div>
-);
-
-export default DesDevSwitch;
