+++ next-docs/pages/gettingStarted.tsx
@@ -1,21 +1,28 @@
 import { useState } from 'react';
-import DesDevSwitch from '../components/gettingStarted/DesDevSwitch';
+import { Tabs } from '@heathmont/moon-core-tw';
 import DesGettingStarted from '../components/gettingStarted/DesGettingStarted';
 import DevGettingStarted from '../components/gettingStarted/DevGettingStarted';
 
 const PageGettingStarted = () => {
-  const [isDesActive, setActive] = useState(true);
-  const setDesActive = () => setActive(true);
-  const setDevActive = () => setActive(false);
+  const [selectedIndex, setSelectedIndex] = useState(0);
+  const handleClick = () => setSelectedIndex(selectedIndex);
   return (
     <>
       <h1 className="text-moon-32 font-medium">Getting started</h1>
-      <DesDevSwitch
-        isDesActive={isDesActive}
-        setDesActive={setDesActive}
-        setDevActive={setDevActive}
-      />
-      {isDesActive ? <DesGettingStarted /> : <DevGettingStarted />}
+      <Tabs selectedIndex={selectedIndex} onChange={handleClick}>
+        <Tabs.Segment size="sm">
+          <Tabs.Pill>I'm a designer</Tabs.Pill>
+          <Tabs.Pill>I'm a developer</Tabs.Pill>
+        </Tabs.Segment>
+        <Tabs.Panels>
+          <Tabs.Panel className="flex flex-col gap-12">
+            <DesGettingStarted />
+          </Tabs.Panel>
+          <Tabs.Panel className="flex flex-col gap-12">
+            <DevGettingStarted />
+          </Tabs.Panel>
+        </Tabs.Panels>
+      </Tabs>
     </>
   );
 };
