--- next-docs/pages/start/reset.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function PageBadge() {
-  return (
-    <>
-      <h1>Reset</h1>
-    </>
-  );
-}
